Added table sports parts
These are parts of various table sports games. The table soccer is for letting one player flick the ball at the goal, across the table. The other player tries to play goalie. The field goal is for paper football. The clips at the bottom of the goals are so that the pieces latch on to my table.